AST to Timeline

+++ lmml/src/timeline.rs
@@ -17,7 +17,7 @@ pub struct Note {
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum NoteType {
-    Single { hz: f64, volume: f32 },
+    Single { hz: f32, volume: f32 },
     Rest,
 }
